switch to 64 bit floating point

+++ src/kmeans.rs
@@ -1,19 +1,19 @@
 /// Compute the SSE of k-means on a 1-dimensional dataset.
-pub fn kmeans_error(k: usize, data: &[f32], max_iters: usize) -> f32 {
+pub fn kmeans_error(k: usize, data: &[f64], max_iters: usize) -> f64 {
     // Create initial centroids
     let mut centroids = initialize_centroids(k, data);
 
     let mut error = 0.0;
-    let mut clusters: Vec<Vec<f32>> = vec![vec![]; k];
+    let mut clusters: Vec<Vec<f64>> = vec![vec![]; k];
     for _ in 0..max_iters {
         error = 0.0;
 
         // Assign each point to a centroid
         for point in data {
             let mut current_centroid = 0;
-            let mut best_dist = f32::MAX;
+            let mut best_dist = f64::MAX;
             for (i, centroid) in centroids.iter().enumerate() {
-                let distance = f32::abs(centroid - point);
+                let distance = f64::abs(centroid - point);
                 if distance < best_dist {
                     current_centroid = i;
                     best_dist = distance;
@@ -24,14 +24,14 @@ pub fn kmeans_error(k: usize, data: &[f32], max_iters: usize) -> f32 {
 
         // Move each centroid closer to the center of the cluster
         for (centroid, cluster) in centroids.iter_mut().zip(clusters.iter()) {
-            let mean = cluster.iter().sum::<f32>() / cluster.len() as f32;
+            let mean = cluster.iter().sum::<f64>() / cluster.len() as f64;
             *centroid = mean;
 
             // Compute the sum of squared error
             error += cluster
                 .iter()
-                .map(|x| f32::abs(x - *centroid).powi(2))
-                .sum::<f32>();
+                .map(|x| f64::abs(x - *centroid).powi(2))
+                .sum::<f64>();
         }
 
         // Reset clusters
@@ -43,7 +43,7 @@ pub fn kmeans_error(k: usize, data: &[f32], max_iters: usize) -> f32 {
     error
 }
 
-pub fn get_clusters(k: usize, data: &[f32], max_iters: usize) -> Vec<Vec<f32>> {
+pub fn get_clusters(k: usize, data: &[f64], max_iters: usize) -> Vec<Vec<f64>> {
     // Create initial centroids
     let mut centroids = initialize_centroids(k, data);
 
@@ -57,9 +57,9 @@ pub fn get_clusters(k: usize, data: &[f32], max_iters: usize) -> Vec<Vec<f32>> {
         // Assign each point to a centroid
         for point in data {
             let mut current_centroid = 0;
-            let mut best_dist = f32::MAX;
+            let mut best_dist = f64::MAX;
             for (i, centroid) in centroids.iter().enumerate() {
-                let distance = f32::abs(centroid - point);
+                let distance = f64::abs(centroid - point);
                 if distance < best_dist {
                     current_centroid = i;
                     best_dist = distance;
@@ -70,7 +70,7 @@ pub fn get_clusters(k: usize, data: &[f32], max_iters: usize) -> Vec<Vec<f32>> {
 
         // Move centroids
         for (centroid, cluster) in centroids.iter_mut().zip(clusters.iter()) {
-            let mean = cluster.iter().sum::<f32>() / cluster.len() as f32;
+            let mean = cluster.iter().sum::<f64>() / cluster.len() as f64;
             *centroid = mean;
         }
     }
@@ -78,7 +78,7 @@ pub fn get_clusters(k: usize, data: &[f32], max_iters: usize) -> Vec<Vec<f32>> {
     clusters
 }
 
-fn initialize_centroids(k: usize, data: &[f32]) -> Vec<f32> {
+fn initialize_centroids(k: usize, data: &[f64]) -> Vec<f64> {
     let mut centroids = Vec::with_capacity(k);
 
     let min = data
@@ -93,9 +93,9 @@ fn initialize_centroids(k: usize, data: &[f32]) -> Vec<f32> {
     // Evenly distribute over [min, max]
 
     for i in 0..k {
-        // let index = (i as f32 * data.len() as f32 / k as f32) as usize;
+        // let index = (i as f64 * data.len() as f64 / k as f64) as usize;
         // centroids.push(data[index]);
-        centroids.push((max - min) / (k as f32 - 1.0) * i as f32 + min);
+        centroids.push((max - min) / (k as f64 - 1.0) * i as f64 + min);
     }
     centroids
 }
